requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
msal>=1.24.0
requests>=2.31.0
